[Parcelize] Don't use lookup predicate for matching parcelable classes
After previous commits, any class referenced by lookup predicate is
considered as dirty for the incremental compilation. Since parcelize
actually doesn't need to find all parcelable classes, but either needs
just to check if something is parcelable, it's better to use
`DeclarationPredicate` for that, so the IC won't be affected.

The need of this change was found thanks to the following test:
`Kapt4AndroidIncrementalWithoutPreciseBackupIT.testAndroidDaggerIC`

^KT-75864