[LL] ContextCollector: dump context only for real source elements
Some fake sources might be traversed and cached before the real element,
so an incorrect result might be returned.
That was the case for enum classes. They have a synthetic super type
reference with the fake (`EnumSuperTypeRef`) source which contains
the class psi. As a result, the incorrect context was captured and
returned for the class

^KT-79653 Fixed